+++ CMakeLists.txt
@@ -15,7 +15,7 @@ if(APPLE)
 endif()
 
 include_directories(include)
-add_executable(screenreclib main.cpp lib/SRDecoder.cpp include/SRDecoder.h lib/SREncoder.cpp include/SREncoder.h lib/SRMediaOutput.cpp include/SRMediaOutput.h lib/SRVideoInput.cpp include/SRVideoInput.h lib/SRAudioInput.cpp include/SRAudioInput.h include/SRTools.h lib/SRInput.cpp include/exceptions/SRExceptions.h lib/SRTranscoder.cpp include/SRTranscoder.h)
+add_executable(screenreclib main.cpp lib/SRDecoder.cpp include/transcoding/SRDecoder.h lib/SREncoder.cpp include/transcoding/SREncoder.h lib/SRMediaOutput.cpp include/muxing/SRMediaOutput.h lib/SRVideoInput.cpp include/demuxing/SRVideoInput.h lib/SRAudioInput.cpp include/demuxing/SRAudioInput.h include/SRTools.h lib/SRInput.cpp include/exceptions/SRExceptions.h lib/SRTranscoder.cpp include/transcoding/SRTranscoder.h)
 
 target_link_libraries(
         screenreclib
