+++ include/transcoding/SREncoder.h
