+++ lib/SRAudioInput.cpp
@@ -2,7 +2,7 @@
 // Created by Giulio Carota on 17/10/21.
 //
 
-#include "SRAudioInput.h"
+#include "demuxing/SRAudioInput.h"
 
 /**
  * the constructor initialize the
