--- lib/SRVideoDecoder.cpp
@@ -1,5 +1,0 @@
-//
-// Created by Giulio Carota on 17/10/21.
-//
-
-#include "../include/SRVideoDecoder.h"
